# Razor Extension Serialization Refactoring