--- a/src/Razor/src/Microsoft.CodeAnalysis.Razor.Workspaces/Serialization/SerializedRazorExtension.cs
+++ b/src/Razor/src/Microsoft.CodeAnalysis.Razor.Workspaces/Serialization/SerializedRazorExtension.cs
@@ -4,7 +4,7 @@
 using System;
 using Microsoft.AspNetCore.Razor.Language;
 
-namespace Microsoft.VisualStudio.LanguageServices.Razor.Serialization
+namespace Microsoft.CodeAnalysis.Razor.Serialization
 {
     internal class SerializedRazorExtension : RazorExtension
     {