… src/Razor/src/Microsoft.AspNetCore.Razor.LanguageServer.Common/Serialization/RazorExtensionJsonConverter.cs
… src/Razor/src/Microsoft.AspNetCore.Razor.LanguageServer.Common/Serialization/SerializedRazorExtension.cs