This isn't really a perf thing, just updating for consistency + removing duplicated files across the project.

|                                                                  Method |     Mean |     Error |    StdDev |  Op/s |   Gen 0 |   Gen 1 |   Gen 2 | Allocated |
|------------------------------------------------------------------------ |---------:|----------:|----------:|------:|--------:|--------:|--------:|----------:|
| 'Before FullProjectSnapshotHandle Serialization' | 3.294 ms | 0.0651 ms | 0.0724 ms | 303.6 | 35.1563 | 15.6250 |   2.81 MB |
| 'After FullProjectSnapshotHandle Serialization' | 3.338 ms | 0.0655 ms | 0.0918 ms | 299.6 | 35.1563 | 15.6250 |   2.81 MB |